🛠 refactor: ts 마이그레이션
