+++ package.json
@@ -4,7 +4,7 @@
   "description": "My Own Blog Project",
   "main": "app.js",
   "scripts": {
-    "start": "NODE_ENV=production tsc-watch --onSuccess 'node ./dist/index.js'",
+    "start": "tsc && NODE_ENV=production node ./dist/index.js",
     "dev": "NODE_ENV=development tsc-watch --onSuccess 'node ./dist/index.js'",
     "build": "tsc"
   },
