+++ src/controllers/index.ts
@@ -1,5 +1,5 @@
 export * from './admin';
-export * from './comment';
+// export * from './comment';
 export * from './post';
 export * from './user';
 export * from './upload';
