+++ src/db/models/comment.ts
