+++ src/routers/apis/admin.ts
@@ -1,11 +1,6 @@
 import { Router } from 'express';
 import { asyncHandler, isAdmin } from '../middlewares';
-import {
-  createCategory,
-  createSubcategory,
-  deleteCategory,
-  deleteSubcategory,
-} from '../../controllers';
+import { createCategory, createSubcategory } from '../../controllers';
 
 const router = Router();
 
@@ -13,10 +8,10 @@ router.use(isAdmin);
 router.post('/categories', asyncHandler(createCategory));
 router.post('/subcategories', asyncHandler(createSubcategory));
 
-router.delete('/categories/:id', asyncHandler(deleteCategory));
-router.delete(
-  '/categories/:categoryId/subcategory',
-  asyncHandler(deleteSubcategory),
-);
+// router.delete('/categories/:id', asyncHandler(deleteCategory));
+// router.delete(
+//   '/categories/:categoryId/subcategory',
+//   asyncHandler(deleteSubcategory),
+// );
 
 export default router;
