+++ src/routers/apis/post.ts
@@ -1,12 +1,12 @@
 import { Router } from 'express';
 import {
   validatePost,
-  validatePostEdit,
+  // validatePostEdit,
   isAdmin,
   asyncHandler,
 } from '../middlewares';
 // import { clearCache } from '../../middlewares/cleanCache';
-import { createPost, updatePost, deletePost, getPost } from '../../controllers';
+import { createPost, getPost } from '../../controllers';
 
 const router = Router();
 
@@ -14,8 +14,8 @@ router.get('/:postId', asyncHandler(getPost));
 router.use(isAdmin);
 router.post('/', validatePost, asyncHandler(createPost));
 
-router.put('/:id', validatePostEdit, asyncHandler(updatePost));
-
-router.delete('/:id', asyncHandler(deletePost));
+// router.put('/:id', validatePostEdit, asyncHandler(updatePost));
+//
+// router.delete('/:id', asyncHandler(deletePost));
 
 export default router;
