+++ src/routers/middlewares/cleanCache.ts
