+++ src/services/index.ts
@@ -1,5 +1,4 @@
 export * from './category';
 export * from './subcategory';
-export * from './comment';
 export * from './post';
 export * from './user';
