+++ src/views/partials/head.pug
@@ -8,6 +8,7 @@ link(
 )
 link(rel='stylesheet' href='/css/bootstrap.css')
 link(rel='stylesheet' href='/css/index.css')
+link(rel='stylesheet' href='/css/show-post.css')
 script(src='/js/bootstrap.js')
 script(src="https://cdn.jsdelivr.net/npm/axios@1.1.2/dist/axios.min.js")
 script(src="https://cdn.jsdelivr.net/npm/showdown@2.1.0/dist/showdown.min.js")
