+++ src/views/posts/show.pug
@@ -33,4 +33,4 @@ block content
         //-                 button.btn.btn-danger#deleteBtn 댓글 지우기
 block script
     script(src="/js/showPost.js")
-    script(src="/js/comment.js")
+    script(src="/js/comment.ts")
